[feature/splash_screen] 스플래시 화면 파일 분리(#16)
