+++ app/src/main/java/com/everymeal/everymeal_android/SplashScreen.kt
@@ -1,0 +1,48 @@
+package com.everymeal.everymeal_android
+
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.everymeal.everymeal_android.ui.theme.EveryMeal_AndroidTheme
+
+@Composable
+fun SplashScreen() {
+    Column(
+        modifier = Modifier.fillMaxSize(),
+        horizontalAlignment = Alignment.CenterHorizontally,
+        verticalArrangement = Arrangement.Center
+    ) {
+        Image(
+            painter = painterResource(id = R.drawable.every_meal_logo_image),
+            contentDescription = stringResource(R.string.app_name),
+            modifier = Modifier.size(80.dp)
+        )
+        Spacer(modifier = Modifier.size(8.dp))
+        Text(
+            text = "everymeal",
+            fontSize = 24.sp,
+            fontWeight = FontWeight.Bold
+        )
+    }
+}
+
+@Preview(showBackground = true)
+@Composable
+fun SplashPreview() {
+    EveryMeal_AndroidTheme {
+        SplashScreen()
+    }
+}
